fix: make getSlug async + util function for retrieving tag suggestions
